unmatched pattern should NOT raise error in bash (it does in zsh)
